# api/API/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de blog du cours Api",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "sandbox-server"
        }
    ],
    "paths": {
        "/animal": {
            "get": {
                "summary": "List all animals",
                "description": "List all animals",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit permettant la pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List all animals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Animal[]",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Animal"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animal/{id}": {
            "get": {
                "summary": "return an animal",
                "description": "return an animal",
                "operationId": "e84f7fbb9a58ad9ffc06280fd53931a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de l'animal à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return an animal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Animal"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animal/create": {
            "post": {
                "summary": "Create an animal",
                "description": "Save animal in DB",
                "operationId": "6656c4c0539c20b0329a777ba83e11db",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Animal à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Charlie"
                                    },
                                    "espece": {
                                        "type": "string",
                                        "example": "Chiens"
                                    },
                                    "couleur": {
                                        "type": "string",
                                        "example": "Blanc et marron"
                                    },
                                    "race": {
                                        "type": "string",
                                        "example": "Welsh Corgi Pembroke"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "taille": {
                                        "type": "double",
                                        "example": 0.5
                                    },
                                    "poids": {
                                        "type": "double",
                                        "example": 10
                                    },
                                    "sexe": {
                                        "type": "bit",
                                        "example": 0
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Très joueur"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://www.centrale-canine.fr/sites/default/files/2019-02/Big%20Woods%20Thomas%20O%27Malley%20%28Johana%20Flink%29.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Animal enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animal/{id}/update": {
            "put": {
                "summary": "Update animal",
                "description": "Update animal in Db",
                "operationId": "9becc98c0e82e95fc87181945fe032e7",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de l'animal à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Nouvelles données de l'animal",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Charlie"
                                    },
                                    "espece": {
                                        "type": "string",
                                        "example": "Chiens"
                                    },
                                    "couleur": {
                                        "type": "string",
                                        "example": "Blanc et marron"
                                    },
                                    "race": {
                                        "type": "string",
                                        "example": "Welsh Corgi Pembroke"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "taille": {
                                        "type": "double",
                                        "example": 0.5
                                    },
                                    "poids": {
                                        "type": "double",
                                        "example": 12
                                    },
                                    "sexe": {
                                        "type": "bit",
                                        "example": 0
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Très joueur"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://www.centrale-canine.fr/sites/default/files/2019-02/Big%20Woods%20Thomas%20O%27Malley%20%28Johana%20Flink%29.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Animal enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animal/{id}/delete": {
            "delete": {
                "summary": "Delete animal",
                "description": "Delete animal in Db",
                "operationId": "102e6a5cf41dafbe072e95225b6587d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de l'animal à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression validée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Animal supprimé"
                            }
                        }
                    }
                }
            }
        },
...
    }
}